Move pass_fulcra_api decorator to utils.py
Move the decorator to cli/utils.py alongside the requires_auth decorator
instead of creating a separate module. This keeps all CLI decorators in
one place without adding an extra file.
