+++ fulcra_api/cli/__init__.py
@@ -4,10 +4,6 @@ import click
 
 from ..core import FulcraAPI
 from ..credentials import FulcraCredentials
-
-# Create a pass decorator for FulcraAPI to enable type hints in subcommands
-pass_fulcra_api = click.make_pass_decorator(FulcraAPI)
-
 from .auth import auth
 from .commands import (
     catalog,
@@ -30,7 +26,7 @@ from .commands import (
 from .data_types import data_type
 from .files import file
 from .tags import tag
-from .utils import ensure_config_directory, load_creds, save_creds
+from .utils import ensure_config_directory, load_creds, save_creds, pass_fulcra_api
 
 
 @click.group()
