+++ fulcra_api/cli/data_types.py
@@ -6,8 +6,7 @@ from uuid import UUID
 import click
 
 from ..core import FulcraAPI
-from . import pass_fulcra_api
-from .utils import requires_auth
+from .utils import pass_fulcra_api, requires_auth
 
 
 @click.group(name="data-type", help="Data type management sub-commands")
