+++ fulcra_api/cli/utils.py
@@ -6,7 +6,11 @@ from functools import wraps
 import click
 import dateparser
 
+from ..core import FulcraAPI
 from ..credentials import FulcraCredentials
+
+# Create a pass decorator for FulcraAPI to enable type hints in subcommands
+pass_fulcra_api = click.make_pass_decorator(FulcraAPI)
 
 CONFIG_PATH = pathlib.Path.home() / ".config" / "fulcra"
 CREDS_FILE = pathlib.Path(CONFIG_PATH / "credentials.json")
